Move everything from the main activity over to a fragment
