+++ app/src/main/res/layout/activity_main.xml
@@ -4,94 +4,18 @@
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#ccdfff"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="andrewbeav.github.io.weathermeme.MainActivity">
 
-    <ImageView
-        android:id="@+id/memeView"
+    <android.support.v4.view.ViewPager
+        android:id="@+id/pager"
         android:layout_width="match_parent"
-        android:layout_height="250dp"
-        android:src="@drawable/main_image"
-    />
-
-
-    <Button
-        android:id="@+id/edit_location_button"
-        android:background="@color/activity_main_button_color"
-        android:textColor="@color/activity_main_button_text_color"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_below="@+id/memeView"
-        android:layout_marginLeft="10dp"
-        android:layout_marginTop="15dp"
-        android:onClick="editLocation"
-        android:text="@string/edit_location_text"
-        android:layout_width="match_parent"
-        android:layout_height="40dp" />
-
-    <Button
-        android:id="@+id/refresh_button"
-        android:background="@color/activity_main_button_color"
-        android:textColor="@color/activity_main_button_text_color"
-        android:layout_marginTop="15dp"
-        android:layout_centerHorizontal="true"
-        android:layout_alignParentBottom="true"
-        android:text="@string/refresh_button_text"
-        android:onClick="refreshWeather"
-        android:layout_width="wrap_content"
-        android:layout_height="40dp" />
-
-    <GridLayout
-        android:id="@+id/weather_info_layout"
-        android:layout_below="@+id/edit_location_button"
-        android:layout_width="match_parent"
-        android:layout_marginTop="10dp"
         android:layout_height="match_parent">
 
-        <TextView
-            android:id="@+id/temp_text"
-            android:text="@string/temp_default"
-            android:textSize="20dp"
-            android:layout_row="0"
-            android:layout_column="0"
-            android:layout_margin="10dp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-        />
-        <TextView
-            android:id="@+id/humidity_text"
-            android:text="@string/humidity_default"
-            android:textSize="20dp"
-            android:layout_margin="10dp"
-            android:layout_row="0"
-            android:layout_column="1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-        />
+    </android.support.v4.view.ViewPager>
 
-        <TextView
-            android:id="@+id/wind_text"
-            android:text="@string/wind_default"
-            android:textSize="20dp"
-            android:layout_row="1"
-            android:layout_column="0"
-            android:layout_margin="10dp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-        />
-        <TextView
-            android:id="@+id/weather_description_text"
-            android:text="@string/weather_description_default"
-            android:textSize="20dp"
-            android:layout_row="1"
-            android:layout_column="1"
-            android:layout_margin="10dp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-        />
-    </GridLayout>
+
 </RelativeLayout>
