+++ app/src/main/res/values/strings.xml
@@ -9,4 +9,7 @@
     <string name="location_text_view_text">Example: Stillwater,Ok</string>
     <string name="submit_location_text">Submit</string>
     <string name="use_current_location_text">Use Current Location</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
